Add xpdf and pdftohtml source code to the project
Current pdf to html conversion softwares(either free or non-free) are not satisfying for two reasons. First, they don't support Chinese characters which is a fatal drawback to me as a Chinese. Second, there are still rooms to improve in many aspects(the quality of resulting HTML of pdftohtml is sometimes too bad to be accepted by user). Third, the conversion tool should work well with my online reader for PDF.
My plan is to write my own pdf-to-html conversion tool based on the code of xpdf and pdftohtml.